+++ tests-out/4.err
@@ -1,0 +1,1 @@
+reverse: input and output file must differ
