+++ tests-out/4.rc
@@ -1,1 +1,1 @@
-0
+1
