--- tests-out/6.other
@@ -1,5 +1,0 @@
-test
-a
-is
-this
-hello
